api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.9.0
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.3
pandas==2.2.0
pydantic==2.5.3
python-multipart==0.0.6
lxml==5.1.0
aiofiles==23.2.1

